% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_local.R
\name{connect_local}
\alias{connect_local}
\title{Open a connection to the local database}
\usage{
connect_local(path = "~/analysis/n2kreport.sqlite")
}
\arguments{
\item{path}{the path to the local database}
}
\description{
Open a connection to the local database
}

