% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_remote.R
\name{connect_remote}
\alias{connect_remote}
\title{Make a read-only connection to the remote database}
\usage{
connect_remote(develop = TRUE)
}
\arguments{
\item{develop}{Use the development database (TRUE) or the production database
(FALSE).}
}
\description{
Make a read-only connection to the remote database
}

