% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_dummy_db.R
\name{create_dummy_db}
\alias{create_dummy_db}
\title{Create a dummy database}
\usage{
create_dummy_db(connection)
}
\arguments{
\item{connection}{A DBI connection to the dummy database. Defaults to an in
memory SQLite database}
}
\description{
This database is used to test several functions. It allows to get the Shiny
app running without connection to the original database. It will not
overwrite existing data
}

