% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_index.R
\name{export_index}
\alias{export_index}
\title{Prepare indices for export}
\usage{
export_index(connection, table)
}
\arguments{
\item{connection}{An open connection to the SQLite database}

\item{table}{the table in which the species are stored}
}
\description{
Prepare indices for export
}

