% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_plot.R
\name{export_plot}
\alias{export_plot}
\title{Export a single plot}
\usage{
export_plot(file, plot, width = 10, height = 10)
}
\arguments{
\item{file}{the name of the output file}

\item{plot}{a ggplot2 object}

\item{width}{the width in cm}

\item{height}{the height in cm}
}
\description{
Export a single plot
}

