% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_plot_zip.R
\name{export_plot_zip}
\alias{export_plot_zip}
\title{Export all indices as plots}
\usage{
export_plot_zip(zipfile, connection, base_size = 12, width = 10,
  height = 10)
}
\arguments{
\item{base_size}{The base font size}

\item{width}{the width in cm}

\item{height}{the height in cm}
}
\description{
Export all indices as plots
}

