% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_index.R
\name{gg_index}
\alias{gg_index}
\title{Create a ggplot2 object on an index}
\usage{
gg_index(index, baseline, backtransform = TRUE, breaks = NULL,
  labels = NULL)
}
\arguments{
\item{index}{A \code{data.frame} with the indices}

\item{baseline}{An optional number}

\item{backtransform}{A logical value indicating is the results should be
displayed in the original scale (\code{TRUE}) or in the log-scale
(\code{FALSE})}
}
\description{
Create a ggplot2 object on an index
}

