% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_not_available.R
\name{gg_not_available}
\alias{gg_not_available}
\title{ggplot object with text "No results available"}
\usage{
gg_not_available()
}
\description{
ggplot object with text "No results available"
}

