% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_index.R
\name{import_index}
\alias{import_index}
\title{import indices per species}
\usage{
import_index(channel)
}
\arguments{
\item{channel}{An open ODBC connection to the database}
}
\description{
import indices per species
}

