% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Final preparation of dataset and plot}
\usage{
plot_index(to.plot, base_size = 12)
}
\arguments{
\item{to.plot}{the dataset to plot}

\item{base_size}{The base font size}
}
\description{
Final preparation of dataset and plot
}

