% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_composite_index.R
\name{read_composite_index}
\alias{read_composite_index}
\title{Read the results for the composite indices}
\usage{
read_composite_index(result.channel)
}
\arguments{
\item{result.channel}{An open ODBC connection to the database}
}
\value{
a \code{tbl_df} object with the data
}
\description{
Read the results for the composite indices
}

