% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_species.R
\name{read_species}
\alias{read_species}
\title{Get a list of the available species}
\usage{
read_species(connection, table, variable)
}
\arguments{
\item{connection}{An open connection to the SQLite database}

\item{table}{the table in which the species are stored}

\item{variable}{the column in which the species are stored}
}
\description{
Get a list of the available species
}

