% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refresh_persisent.R
\name{refresh_persisent}
\alias{refresh_persisent}
\title{Refresh the persisent storage}
\usage{
refresh_persisent(remote.db, local.db)
}
\arguments{
\item{remote.db}{an open ODBC connection to the SQL Server database}

\item{local.db}{a connection to a local database}
}
\description{
The original data is stored in an SQL Server database. This function caches
the relevant data from the SQL Server localy.
}

