% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_reference.R
\name{set_reference}
\alias{set_reference}
\title{Use the first value of Period as a reference}
\usage{
set_reference(index)
}
\arguments{
\item{index}{a data.frame}
}
\description{
Use the first value of Period as a reference
}

