% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/start_n2k_composite_index.R
\name{start_n2k_composite_index}
\alias{start_n2k_composite_index}
\title{Start the n2k_composite_index Shiny app}
\usage{
start_n2k_composite_index(local.path = "~/analysis/n2kreport.sqlite")
}
\arguments{
\item{local.path}{the path to store the local datawarehouse}
}
\description{
The n2k_composite_index app displays composite indices from the Natura 2000 monitoring
}

