% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_plot_zip.R
\name{export_plot_zip}
\alias{export_plot_zip}
\title{Export all indices as plots}
\usage{
export_plot_zip(zipfile, connection, table, variable, base_size = 12,
  width = 10, height = 10, reference = FALSE)
}
\arguments{
\item{zipfile}{The pathname of the zip file: tilde expansion (see
    \code{\link{path.expand}}) will be performed.}

\item{connection}{An open connection to the SQLite database}

\item{table}{the table in which the species are stored}

\item{variable}{the column in which the species are stored}

\item{base_size}{The base font size}

\item{width}{the width in cm}

\item{height}{the height in cm}

\item{reference}{Use the first index as reference. Default = FALSE}
}
\description{
Export all indices as plots
}

