% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_index.R
\name{read_index}
\alias{read_index}
\title{read the required indices}
\usage{
read_index(connection, species, frequency)
}
\arguments{
\item{connection}{An open connection to the SQLite database}

\item{species}{the ID of the relevant species}

\item{frequency}{Return the yearly indices or per cycle}
}
\description{
read the required indices
}

