% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_species.R
\name{read_species}
\alias{read_species}
\title{Get a list of the available species}
\usage{
read_species(connection)
}
\arguments{
\item{connection}{An open connection to the SQLite database}
}
\description{
Get a list of the available species
}

